Added picker text view
Picker text view is a subclass of UITextView that can have a
placeholder and a border similar to UITextField. Additionally, picker
text view handles editing events when the input view of the text view
is no longer the system keyboard by preventing copy and paste or select
all and removing the cursor.